crach to fix

+++ libwdiv/include/compiler.hpp
@@ -239,21 +239,21 @@ private:
   void emitGosubTo(int targetOffset);
   void patchJumpTo(int operandOffset, int targetOffset);
 
-  void handle_assignment(uint8 getOp, uint8 setOp, int arg, bool canAssign);
+  void handle_assignment(uint16 getOp, uint16 setOp, int arg, bool canAssign);
 
   void prefixIncrement(bool canAssign);
   void prefixDecrement(bool canAssign);
 
   // Variables
-  uint8 identifierConstant(Token &name);
+  uint16 identifierConstant(Token &name);
   void namedVariable(Token &name, bool canAssign);
-  void defineVariable(uint8 global);
+  void defineVariable(uint16 global);
   void declareVariable();
   void addLocal(Token &name);
   int resolveLocal(Token &name);
   void markInitialized();
 
-  uint8 argumentList();
+  uint16 argumentList();
 
   void compileFunction(Function *func, bool isProcess);
   void compileProcess(const std::string &name);
